backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.6.0
pandas>=2.2.0
openpyxl==3.1.2
xlrd==2.0.1
alembic==1.12.1
python-dateutil==2.8.2
python-docx==1.1.0

